Image/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.9.2
huggingface_hub==0.26.2
langchain==0.0.190
python-dotenv==1.0.1
jinja2==3.1.4
requests==2.32.3
pymysql==1.1.1